add elmentwise multiplication
