+++ stk/ops/__init__.py
@@ -1,2 +1,3 @@
-from stk.ops.linear_ops import dds, dsd, sdd, mul
+from stk.ops.linear_ops import dds, dsd, sdd
 from stk.ops.matrix_ops import ones_like, row_indices, sum, to_dense, to_sparse
+from stk.ops.eltwise_ops import mul
