+++ stk/ops/eltwise_ops_test.py
@@ -15,8 +15,6 @@ def allclose(x, y, pct=0.25):
         return False
     return True
 
-
-# element-wise mul only uses (m, n)
 _MUL_MATRIX_SIZES = (
     (128, 128, 0.0),
     (256, 256, 0.5),
@@ -34,31 +32,17 @@ _MUL_MATRIX_SIZES = (
     (1024, 1024, 0.75),
 )
 
-
-_TRANSPOSE = (
-    (False, False),
-    (False, True),
-    (True, False),
-    (True, True),
-)
-
-# both must be true 
-_MUL_TRANSPOSE = (
-    (False, False),
-    (True, True),
-)
-
 _DTYPE = (
     torch.float16, torch.bfloat16
 )
 
-def _generate_testcases(_SIZES, _CUSTOM_TRANSPOSE):
-    testcases = itertools.product(_SIZES, _CUSTOM_TRANSPOSE, _DTYPE)
-    testcases = [(*size, *trans, 128, dtype) for 
-        (size, trans, dtype) in testcases]
+def _generate_testcases():
+    testcases = itertools.product(_MUL_MATRIX_SIZES, _DTYPE)
+    testcases = [(*size, 128, dtype) for 
+        (size, dtype) in testcases]
     return testcases
 
-_MUL_TEST = _generate_testcases(_MUL_MATRIX_SIZES, _MUL_TRANSPOSE)
+_MUL_TEST = _generate_testcases()
 
 def _dense_and_sparse_like(x, std=0.1):
 
@@ -82,28 +66,19 @@ def _dense_and_sparse(rows, cols, sparsity, blocking, dtype, std=0.1):
     return (dense.to(cuda_device).requires_grad_(True),
             sparse.to(cuda_device).requires_grad_(True))
 
-def _with_transpose(op, a, b, trans_a, trans_b):
-    a = a.t() if trans_a else a
-    b = b.t() if trans_b else b
-    return op(a, b)
-
-def _mask(x, mask):
-    mask = stk.ops.to_dense(stk.ops.ones_like(mask))
-    return x * mask
 
 
 class LinearOpsTest(parameterized.TestCase):
 
     @parameterized.parameters(_MUL_TEST)
-    def testElemwiseMulWithLike(self, m, n, sparsity, trans_a, trans_b, blocking, dtype):
+    def testElemwiseMulWithLike(self, m, n, sparsity, blocking, dtype):
 
-        a_shape = (n, m) if trans_a else (m, n)
         # Common mask used to ensure same topology for a and b
-        a_dense, a = _dense_and_sparse(*a_shape, sparsity, blocking, dtype)
+        a_dense, a = _dense_and_sparse(m, n, sparsity, blocking, dtype)
         b_dense, b = _dense_and_sparse_like(a)
 
-        out = _with_transpose(stk.ops.mul, a, b, trans_a, trans_b)
-        expected_out = _with_transpose(torch.mul, a_dense, b_dense, trans_a, trans_b)
+        out = stk.ops.mul(a, b)
+        expected_out = torch.mul(a_dense, b_dense)
 
         # Compute the gradients w.r.t. the inputs.
         expected_out.sum().backward()
